文章服务-定向更新文章详情

+++ heima-leadnews/heima-leadnews-model/src/main/java/com/heima/model/wemedia/dtos/WmNewsDto.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Data
 public class WmNewsDto {
     
-    private Integer id;
+    private Long id;
      /**
      * 标题
      */
@@ -17,7 +17,7 @@ public class WmNewsDto {
      /**
      * 频道id
      */
-    private Integer channelId;
+    private Long channelId;
      /**
      * 标签
      */
@@ -33,12 +33,12 @@ public class WmNewsDto {
      /**
      * 文章封面类型  0无图  1单图  3多图  -1自动
      */
-    private Short type;
+    private Integer type;
 
      /**
      * 状态 提交为1  草稿为0
      */
-    private Short status;
+    private Integer status;
      
      /**
      * 封面图片列表 多张图以逗号隔开
@@ -46,7 +46,7 @@ public class WmNewsDto {
     private List<String> images;
 
     //0下架， 1上架
-    private Short enable;
+    private Integer enable;
 
 
 }
